requirements.txt
APScheduler==3.8.1
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==5.0.0
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
decorator==5.1.1
google-auth==2.5.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
gspread==5.1.1
gspread-pandas==3.0.4
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
moodlepy==0.23.4
numpy==1.21.5
oauthlib==3.1.1
pandas==1.3.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
zipp==3.7.0
